Rename auth.ts to user.ts
